return a value in the currencies base unit
sometimes it is beneficial to return the base unit value of a currency rather than the subunit value. this commit allows us to do that by passing a boolean to the `getAmount` method. this implementation will not break BC, and users can continue to use the `getAmount` method as before.

in fact, the PHP number formatter doesn't do any of this math for you, which is why you needed to include it in your `IntlFormatter`. This commit allows us to move that responsibility from the formatter to the `Money` class.

if the mixed return types bothers you, I could also extract this out into a separate method, something like `getConvertedAmount()`