add a changeling
a file to keep track of all notable changes so users will have an easy
place to go to see any updates.

it should track things that are `Added`, `Deprecated`, `Fixed`,
`Removed`, and any `Security` issues.

Obviously it would be a little difficult to go back and add this info
for all previous releases, but with this commit at least we have a
history of versions and their release dates, and a place to document
changes as the package moves forward.